backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.0.5
Werkzeug==3.0.3

# Core HTTP & parsing
requests==2.32.5
beautifulsoup4==4.13.5
lxml==4.9.3
python-dateutil==2.8.2

# AI / LLM
google-generativeai==0.3.2

# Scraping / automation
selenium==4.15.0
newspaper3k==0.2.8

# Env & utils
python-dotenv==1.0.1
reportlab==3.6.12

# Friend's module deps (social/youth analysis)
praw==7.7.1
tweepy==4.12.1
google-api-python-client==2.149.0
textblob==0.17.1
vaderSentiment==3.3.2
schedule==1.2.0
redis==5.2.0
celery==5.4.0
